Move classes where it belongs.
